requirements.txt
asgiref==3.7.2
dj-database-url==2.0.0
Django==4.2.2
django-ckeditor==6.6.1
django-js-asset==2.0.0
gunicorn==20.1.0
Pillow==9.5.0
psycopg2-binary==2.9.6
sqlparse==0.4.4
typing_extensions==4.7.1
